Refactor quadrature integration
- Split VSO quad into separate compilation unit
- Add rotation matrix from x,y,z to u,v,w
- Update Python notebooks for VSO tests
