+++ swig/simulation/CMakeLists.txt
@@ -13,7 +13,7 @@
 # WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
 # A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 
-set(CALIN_SWIG_FILES atmosphere.i pmt.i vs_optics.i #tracker.i 
+set(CALIN_SWIG_FILES atmosphere.i pmt.i vs_optics.i tracker.i 
 	detector_efficiency.i)
 set(CALIN_SWIG_TARGET simulation)
